pyproject.toml
[project]
name = "task4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.6.1",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
    "tools>=1.0.13",
]
